Unified version of code (works with the 3 radars)

+++ triprexPro/writeData.py
@@ -19,7 +19,7 @@ def createTimeDimension(rootgrpOut, timeRef):
 
    try:
       rootgrpOut.createDimension('time', None)
-      time_ref = rootgrpOut.createVariable('time', np.int64, ('time',))
+      time_ref = rootgrpOut.createVariable('time', np.float64, ('time',))
       time_ref[:] = timeRef
       time_ref = dataAttribute.timeAttributes(time_ref)
       return time_ref
